Add support for linear, radial and sweep gradients (#893)
Superserdes #885 to reduce the clutter of commits I produced. 😅 

A first version of linear, radial and sweep gradients. Has support for
all important attributes on gradients as well as arbitrary affine
transforms.

As mentioned before, performance has not been my major focus for this,
but instead correctness. There probably are some easy (and some harder)
approaches for improving performance, if you have thoughts on this feel
free to share them, but for now I would like to land this more or less
in the current shape and defer optimizations to later.