Alter winit usage
Now, only the method that needs the winit Window gets it rather
than trying to store it locally when a feature is enabled. This
also lets the people embedding this library pass in the reference
to the window rather than any sort of window construction being
needed.

This does mean that the correct window must be passed in.