Refactor the test infrastructure for sparse strips (#925)
This PR proposes refactoring the test infrastructure for sparse strips
rendering. Note that it's not ready for careful review yet as there's
still a bit missing (for example, I haven't moved the snapshot tests yet
to avoid trouble with git LFS), but feedback on the overall idea in the
meanwhile would definitely be appreciated (or we just discuss it on
Wednesday).

To summarize, the main change is that the tests do not live in
`vello_cpu` anymore, but instead in a custom crate `vello_sparse_tests`
(I couldn't take `vello_tests` because that crate already exists). In
addition to that, I created a new `vello_macros` crate which contains a
custom macro that we use to define tests. The basic idea of how this
works is that we annotate each test with this macro (and define some
paramters like the height, width, etc.), and then the macro will
automatically generate different versions of the same tests at runtime.
I think this gives us some nice improvements now, and will gives us a
lot of quality-of-life improvements even further down the road:

1) We do not need to always call the `get_ctx` and `check_ref` method,
but each method is instead supplied the render context as a parameter,
saving some boilerplate code.

2) Right now, our rendering pipeline is based solely on u8/16. Soon
(hopefully in the next 1-2 weeks), I will start working on a version of
fine rasterization that is based on f32 instead and should serve as the
gold standard for correctness (the reference images will be based on
that version of the pipeline). By having such a macro for our tests,
once we get to that point we can simply adapt the macro to generate
versions of the test based on u8 AND f32 and can compare them to the
same reference image (for the u8 version, we will obviously need some
wiggle room and allow discrepancies of +-1/2). But by having this
design, we can simply define one version of the test and automatically
test it against both versions of the pipeline, which I think is
incredibly useful.

3) The above will become even more relevant once we get to SIMD, where
we ideally also reuse each test to run it against that specific SIMD
pipeline, which is very easy to do if every test uses that macro. A
great thing is that we can also conditionally include/exclude tests
based on the target feature we detect, so that we only run SSE tests on
a processor supporting SSE, while we also include AVX2 and AVX512
versions of the tests if supported by the architecture.

4) Finally, by extracting the tests into a custom crate, once we have a
render trait that is implemented by both, vello_hybrid and vello_cpu, it
should be possible to change all occurances of `&mut RenderContext` to
that trait, which should allow us to run those same tests against
vello_hyrbid (again, with some wiggle room for small discrepancies),
which I imagine would be very useful as well.

We can add lots of configuration options, such as for example being able
to temporarily disable specific kinds of tests via environment variables
or completely disabling tests for one renderer if they involve features
that are not supported by both.