Use `peniko::Image` for the image `PaintType`
This is more or less a continuation of
https://github.com/linebender/vello/pull/944, which enabled more easily
using ecosystem types.

Note `vello_api` allows constructing a `Pixmap` from PNG data, but this
PR does not currently implement the same for a `peniko::Image`.

Further, because `Pixmap` holds premultiplied RGBA8, it is not trivial
to convert a `Pixmap` to a `peniko::Image`, as the latter for the moment
only holds regular RGBA8.

For the moment, this does the premultiplication during sampling for fine
rasterization. That might not be the best option, especially if at some
point in the future `peniko::Image` supports more formats than just
RGBA8.