Add f32 fine-rasterization pipeline (#967)
Hopefully the last time we have to regenerate all images, but I can't
promise it. :)

Overall, while I don't think the `FineType` is very polished it, as you
can see it does work pretty well and makes it possible to define each
pipeline once for both, u8 and f32 without having to duplicate code. I
am optimistic that this design could be further evolved to also support
SIMD, though that remains to be seen.

I would also like to test whether we can make do with a default
threshold of 0 for the hybrid renderer, but since the hybrid renderer
currently converts colors to u8 (and then back to f32 on the GPU), this
currently is not possible. If anyone has a git diff of changes I need to
apply to get that to work, I'm happy to try it.