Refactor the benchmarking harness for fine rasterization (#999)
On the one hand, similarly to our current test suite, this PR introduces
a macro that allows generating benchmarks for specific fine
rasterization types on the fly. On the other hand, I also added many
other different benchmarks which _should_ cover every important aspect
of fine rasterization.

I will use this benchmark suite as the basis for my optimizations, and
by the end of my work we will hopefully see great improvements in those!
:)

(Merging this is not super urgent as it is relatively self-contained and
I can just branch off from here, so it can just happen after RustWeek.)