Fix wrong issuing of `pop_clip` command (#1033)
The basic problem was that for the wide tile (1, 0), the first clip path
would trigger a `push_clip`, but the second one would not (since the
strip of the second clip path does not lie on that wide tile). However,
we did trigger a `pop_clip` command. This patch seems to fix the issue.

Fixes #1032