Fix panic in nested clip paths (#1035)
This is an attempt at fixing #1034, and is based on #1033, which should
be merged first. I am not 100% confident that this is the correct
solution, so a careful review would be appreciated.

The problem is as follows: The first clip path is a very small
rectangle, which means that the clip bbox will only cover the wile (1,
0). The second clip path is much bigger and starts at the wide tile (0,
0).

Because of this, once we pop the second clip path, the assignment
```rs
let mut x1 = x0 + strip_width;
```

will end up yielding 4 (since the second, bigger clip path starts at x =
0, and it's strip will be 4 wide). However, since the second clip path
starts outside of the clip bbox,

```rs
let clip_x = clip_bbox.x0() * WideTile::WIDTH;
if clip_x > x {
    col += u32::from(clip_x - x);
    x1 = clip_x.max(x1);
}
```

The `x` value will be adjusted to 256 (since the clip bbox starts at the
first wide tile in the x direction), but note that our x1 is still 4!
Because of this

```rs
 let x2 = next_strip.x;
let clipped_x2 = x2.min((cur_wtile_x + 1) * WideTile::WIDTH);
let width = clipped_x2.saturating_sub(x1);
```

the width calculated here will be much bigger than 256, and thus lead to
a panic.