Improve encoding of gradients (#1023)
This PR makes some changes to the way gradients are encoded. The
motivation for this is two-fold:
- Improve the overall performance of gradient rendering.
- Make gradient rendering more SIMD-friendly.

Overall, the PR contains four changes in total:
- The encoding has been adapted so that the parametric variable `t` for
each gradient is always between 0.0 and 1.0. This will make it easier to
implement an efficient version of the `extend` method. Right now, I
still left the while loop, but this can now be replaced with `(val -
val.floor()).fract()`, which is much more SIMD-friendly.
- The encoding of linear gradients has been simplified, such that the
calculation of the current position along the gradient line is basically
a no-op, giving much better performance.
- Encoding of radial gradients has been completely rewritten.
Previously, we would use a quadratic solver to find the parametric value
`t` for a certain position, which had some issues and also was not
really SIMDifiable. Now, I have instead rewritten it to use the same
approach that is used by Skia, which is faster and will be much faster
once we have SIMD, since the approach is very SIMD-friendly.
- It fixes a bug where the offset for sampling at the pixel center was
applied wrongly.

I _think_ this is pretty much as good as encoding will get, so while I
don't want to make promises, I expect this to be the final form of
gradient encoding and don't think it will change significantly in the
future, meaning that it should also be safe to start implementing this
for the GPU (do note that this only applies to the encoding itself and
not for example the `cur_pos` function, that one will probably change
because the current approach with using `Option` for radial gradients
won't work for SIMD.