Add support for multi-threaded fine rasterization (#1044)
This PR adds support for multi-threaded fine rasterization for
`vello_cpu`. See here for a small demo, as you can see, we get
near-linear speedup for large scenes where fine rasterization is the
main bottleneck (note the frame count in the top right):


https://github.com/user-attachments/assets/bc1daab3-f1cf-48a7-b1bb-0c59cfd135e5

I also wanted to make sure that there is no significant performance
impact from doing the splitting up into regions. Because of this, I
rendered two simple scenes (with just a single rect) multiple times and
then compared the timings:

10000 iterations to 600x600:

Before: 1.106366416s
After: 1.16696825s

1000 iterations to 3840x2160:

Before: 2.161953042s
After: 2.226217167s

So while there is a small slowdown, I don't think it's enough to justify
creating a separate path for the single thread case. It's also a bit
hard to test this automatically since it's based on feature flags, but I
ran the tests manually and the diverging code is small enough that I
think it should be manageable, at least for now. If making
strip-rendering multi-threaded turns out to be more complicated perhaps
we have to figure something out, though.