Add initial support for full multi-threaded rendering (#1047)
# Overview

This PR is a first attempt at adding multi-threading support across the
whole pipeline. The emphasis is on “first attempt”, because the current
state is far from ideal, and I will definitely be refining and improving
it in the near future, but I think it’s good enough to be merged as a
first version, as it also puts many of the necessary building blocks in
place.

From my initial benchmarks, the gains are not impressive when drawing
many simple shapes (in my case rectangles, especially small ones), with
8 threads I’m only getting a speedup for 2x-3x, for very small
rectangles (8x8), multithreading is even slower. I will be doing more
extensive profiling to understand the source of those slowdowns.
However, as soon as more complex geometry like curves are involved, I’m
currently seeing speedups ranging from 5x-6x, which I would say is at
least a good start. There are some obvious things that need to be
optimized (coarse rasterization currently always happens on the main
thread, there is a lot of cloning happening to pass information around
between threads), so hopefully there is room for improvement in the
future. There's also some parts of the code I wouldn't exactly call
beautiful (like the way we handle sending alphas from worker threads)
and will probably rethink in the future.

I had to make some small changes to `vello_common`, but other than that
nearly everything is contained in `vello_cpu`. Perhaps some of the
abstractions I introduced (like the `StripGenerator`) could later on be
moved to `vello_common` and be reused by `vello_hybrid`, but I'm not
sure how useful they would be.

# Changes
Most of the code has been commented, but to provide a small summary of
the main changes:
- In addition to holding an index to an alpha buffer, alpha strips also
hold an identifier that identifies the thread the alpha buffer belongs
to.
- All of the tile/strip generation code has been encapsulated into a
`StripGenerator` struct, that provide a callback to the generated strips
which can then be processed by clients.
- A `Dispatcher` trait is introduced, that has a very similar interface
to the current render context, but explicitly takes a reference to all
the arguments it needs. The main motivation for this is that
multi-threaded and single-threaded rendering are different enough that
it's not really possible to implement them in the same struct (for
example, for multi-threaded rendering we need an owned version of its
arguments, while for single-threaded rendering this would be needless).
In addition, by abstracting this away behind the `Dispatcher` trait,
even if you enable the `multithreading` feature, you can still opt to
just using the single-threaded mode without any overhead.

Conceptually, the current way multi-threading is not too complex: The
main thread sends paths to worker threads, which will then turn the path
into a strips representation. Those worker threads then send those
strips back to the main thread. Order is obviously important here, which
is why each operation is assigned a `cmd_idx`, and when reading from the
ordered channel, the results will be received in the correct order by
the main thread. Then, we simply do coarse rasterization on the main
thread and finally multi-threaded fine rasterization, as it already
exists on main.

An important detail is that we also do batching of tasks: If we draw a
lot of small paths, sending one path per thread would lead to an
incredible overhead due to frequent context switching. Because of this,
we try to very roughly estimate how long it will take to process the
path, and based on that batch multiple paths and then send them all
together to one thread. The current heuristic seems to be working at
least okay, but when I get to fine-tuning I will definitely do more
careful profiling for that to see whether it can be improved.

---------

Co-authored-by: Taj Pereira <taj@canva.com>