Make glyph rendering an optional feature (#1060)
Unfortunately I'm stuck because just enabling the "std" feature will
also enable the "skrifa/std" feature, and because of this skrifa will be
pulled in as a dependency even though the text feature is disabled. I'm
not sure how to best solve this without adding two
`skrifa-std`/`skrifa-libm` features that need to be enabled manually by
the user. Suggestions are welcome.

The main motivation is that it would allow using the renderer in
contexts where text rendering is already implemented with a different
method, without having to pull in skrifa as a dependency (e.g. resvg).