fix: sparse_strips scalar WASM build containing SIMD instructions (#1071)
### Context

This is a bit of a subtle fix that follows
https://github.com/raphlinus/fearless_simd/pull/16

The important excerpts from the WebAssembly spec.

> "[Modules](https://www.w3.org/TR/wasm-core-2/index.html#syntax-module)
are valid when all the components they contain are valid."
> ~[link](https://www.w3.org/TR/wasm-core-2/#modules%E2%91%A2)

Thus, for a browser (or WebAssembly runtime) that doesn't support SIMD,
the module validation check would fail because:
- `v128` type would be invalid and wouldn't exist in that WebAssembly
context.
- `expr` must also be valid, so instructions over f32x4 would fail
validation.

###  Why

Prior to the fearless_simd fix, it was possible via `target_feature` to
accidentally leak SIMD instructions into the binary. Even if these
instructions are never executed, their mere presence would result in the
WASM binary failing to validate.

### Changes

This PR was implemented in a test-driven style.

First I implemented the test:
- The test only runs if the `simd128` target feature is not requested in
the rustc flags.
- The test self-introspects the test wasm binary and validates the
binary against a validator that doesn't contain SIMD support.

This test failed (because of the leaking SIMD instructions). Hence, on a
browser that didn't support SIMD we would crash. (Not good for consumers
of vello using WASM asking for a scalar binary).

Example test failure:

<img width="1022" alt="Screenshot 2025-06-27 at 3 21 59 pm"
src="https://github.com/user-attachments/assets/fe886294-5cb1-40ea-b97d-0f7b10ce62b5"
/>

Then I bumped up fearless_simd, and the test passes!

### Test plan

This PR is effectively a test to prevent regressions.