Rewrite `vello_cpu` to be more SIMD-friendly (#1078)
This is a first attempt at rewriting `vello_cpu` to be more
SIMD-friendly.

As can be seen in the code, there are still quite a few TODOs left. The
usage of SIMD types is also not very consistent (in some cases we are
using f32x16, in some cases we are limited to f32x4), but it should be a
good enough basis for any platform that provides 128-bit SIMD
capabilities, which I think is a good start.

This PR should fix the performance regression that was introduced when
adding the f32 pipeline. There are certain parts of the pipeline (namely
opaque filling and alpha compositing) which only marginally improve as
part of this PR on NEON because it already auto-vectorized very well
before, but other parts like gradients, blurred rects and images
(especially images with bilinear and bicubic filtering) are _much_
faster now.

It also incorporates the gradient LUT change from a previous PR.

I've adjusted a gradient test case so that it doesn't trigger the
floating point impreciseness we have already had before.

---------

Co-authored-by: Andrew Jakubowicz <ajakubowicz@canva.com>