Fix panic in clipping (#1100)
The problem was that when intersecting two clipping boxes, it can happen
that x0/y0 is actually bigger than x1/y1 Therefore, when checking
whether to abort, we should use `>=` rather than `==`, since it's
possible that we already exceeded it in the first iteration.

I tested this fix with 600+ PDFs and it doesn't cause any regressions
there, so I feel pretty certain this is the right fix.