Optimize blending with default blend mode (#1106)
While playing with a profiler and blitz, I noticed that a
non-insignificant amount of time is going into compositing/blending,
even though no special blend modes are used. For the default blend mode,
we can simply use the fast path of doing alpha compositing instead of
taking the slow route.