Two improvements to gradient handling. (#1128)
This PR addresses the issues in #1124 and supersedes #1125.

There are two independent issues that were going on.

Firstly, there was an off-by-one error in our scaling factor for the t
values. If our LUT size is 512, we want the values to be scaled between
0 and 511, and not 0 and 512. I actually already considered this when
calculating the scale factor used in fine rasterization, but when
computing the actual look-up table, we were using the wrong scale value.
Unfortunately, this required regenerating nearly all gradient reference
images, but I think that this should be more correct now (and we have an
additional test case now that verifies this).

Secondly, we based the calculation of the lut size solely on the number
of stops, but we didn't consider the position of the stops. If two stops
are positioned very close to each other, we might have to increase the
resolution. I still hardcoded a maximum limit of 4096 because I don't
think it makes sense to go arbitrarily high there, but hopefully this
should be good enough for all intents and purposes.