Improvements to multi-threaded path rendering (#1133)
This PR contains a number of improvements to multi-threaded rendering,
including:

-  A much simpler and more effective cost estimation for path rendering.
- An optimization to prevent excessive allocations in case we are
drawing many small paths.
- Splitting up the code into multiple files.
- Reusing of the alpha buffers.
- Adding more doc comments.

I have run a couple more experiments with regards to multi-threading,
including:
- Doing coarse rasterization on worker threads by putting it behind a
mutex.
- Doing coarse rasterization completely in parallel by splitting the
canvas into "bands".

Unfortunately, at least from the experiments I've made so far, it only
made things worse. I will do more experiments in that direction, but
this PR contains a number of improvements that should be possible to
land as is, and have a positive effect on performance (validated via
benchmarking).