Alternative API for image overrides (#1161)
Fixes #664

Supersedes (and integrates) #1052

The proposal in #1052 is an ergonomics improvement, but has several
shortfalls. In particular, it required special case methods for every
API which used an external texture. As these only had partial coverage,
you couldn't use an external texture to render text (or any stroked
path).

This is a much smaller proposal based on the core of that PR, to instead
return a (fake) Peniko `Image`, which is then compatible with every
Scene method which would draw an image.

---------

Signed-off-by: Nico Burns <nico@nicoburns.com>
Co-authored-by: Nico Burns <nico@nicoburns.com>