Force the user to specify the render mode in the beginning (#1157)
I think that's reasonable to do, one the one hand it would allow us
making decisions based on the render mode in earlier parts of the
pipeline (though I'm unsure we will ever need it), but more importantly,
it allows us to get the render mode while we are still processing
commands.

For example, up until now, when rendering COLR glyphs, we always used
`OptimizeQuality`, but now we can use whatever mode the user selected.
Another motivation is that for SVG rendering, there are various
situations where we wil need to create sub pixmaps (for rendering
patterns or masks), and it would be good if we can use the same
rendering mode for those.

---------

Co-authored-by: Tom Churchman <thomas@churchman.nl>