Adapt hack for tolerance of stroke expansion (#977)
This was needed in the initial prototype for `vello_cpu` where we were
using the stroke expansion implementation of the GPU stroke expansion
paper, which approximates the stroke by line segments. We had to adjust
the tolerance here, because if you draw a small stroke which is then
scaled up, you can very clearly notice the line approximation.

However, the current (and future) kurbo stroke do cubic-to-cubic stroke
expansion, so I don't think this should be necessary here?