Fix examples crashing when resizing to zero. (#1182)
When a window is resized such that one of its dimensions is zero, the
examples will crash. That's because resizing a `wgpu` surface in such a
way is an illegal operation.

This PR adds a new `valid_surface` flag to track whether the current
surface can actually be drawn on. More serious applications should
delete the surface for the duration, as seen in e.g.
[xilem#1371](https://github.com/linebender/xilem/pull/1371), instead of
tracking the legality with a flag. However, for the examples here,
solving this with a flag was a path of lower resistance and good enough
for examples.

Fixes #371