Store data about filled areas in a strip (#1206)
Currently, in a strip we store the coarse winding number that was
determined while generating it. The thing is that in later part of the
stages, we don't care about the actual winding number but just whether
the area before a strip is filled. Because just the winding number
doesn't give us that information, we need to carry the fill rule around
into every part of the code.

Therefore, I propose changing the representation of a strip from storing
its coarse winding number to just storing a boolean flag that indicates
whether the gap on the left of the strip should be filled. In doing so,
`Vec<Strip>` contains all the information to unambiguously represent the
rendered shape, saving us a lot of work later on.