Fix an bug with clipping of sparse fills (#1194)
The basic problem is that as an optimization the `push_clip` function
does not create `push_buf` commands for any wide tiles that would be
completely filled. However, in `pop_clip` it can happen that due to the
clamping of `x1`, we end up pushing a `ClipFill` command across the
whole wide tile width. I don't think that this is the cleanest solution,
but unfortunately I haven't come up with a different method that doesn't
cause regressions in other cases. This fix works and doesn't cause any
regressions in my integration tests.

Fixes #1186.