Separate storage of strips and alphas from strip generator (#1208)
I based this on top of #1206 to avoid merge conflicts.

Currently, the strip generator also stores the strip and alpha buffers.
This is fine if we can assume that there only is one single buffer for
those globally, but this has already been proven to be a bit awkard once
recordings were introduced. In addition to that, it also makes things
more difficult for the new clipping PR as well as an optimization I want
to implement for multi-threaded rendering.

Therefore, this PR separates them into two different objects by
introducing a new `StripStorage` and forcing the user to explicitly pass
a `StripStorage` that should be used as the basis for the generation
when interacting with the strip generator.