[sparse_strips] fix out of bounds recording crash (#1222)
### Context

Previously, recording crashes when executing strips that are out of
bounds.


https://github.com/user-attachments/assets/f3409cb4-b1b2-45a9-8703-b157ac3ad38d

After this PR – out of bounds strips are handled:



https://github.com/user-attachments/assets/6fe68444-157c-4fca-b84f-7f76952348b5



### Why?

`execute_recording`:
- Iterates over the recording commands, unconditionally calling
`process_geometry_command` on the strips.
- Except – for commands which are fully offscreen there are no strips to
process.

Previously this would cause a crash. The new behavior is we handle an
empty count of strips – meaning there is nothing to render for this
command.

### Test plan

Added a test which renders completely outside the viewport. It used to
crash and no longer crashes. I also tested manually.