fix transform handling in sparse_strips recording API (#1217)
### Context

I was investigating using the recording API ran into a rough edge where
the transform was not correctly captured by the recording.

This PR enforces that transforms are handled in the recording, and adds
more tests to make the recorder API usage clearer.

## Change


https://github.com/linebender/vello/blob/418803ed8420f569b0687b72d5d33d363b90c3d5/sparse_strips/vello_common/src/recording.rs#L426-L430


https://github.com/linebender/vello/blob/418803ed8420f569b0687b72d5d33d363b90c3d5/sparse_strips/vello_common/src/recording.rs#L364-L368

- `Recorder::glyph_run` used `self.recording.transform` eagerly, without
the `recording.transform` ever being set. This resulted in the glyph_run
unconditionally using `Affine::IDENTITY`. This is fixed in this PR by
setting the transform eagerly whenever `set_transform` is called on the
recorder.


However, glyph_run's transform could still be wrong if a `set_transform`
was not within the recording. For example the following code would set a
transform on the rendering context, but then the glyph_run would be
built at `Affine::IDENTITY`:

```rs
    ctx.set_transform(Affine::translate((0., f64::from(font_size))));

    let mut recording = Recording::new();
    ctx.record(&mut recording, |ctx| {
        ctx.set_paint(REBECCA_PURPLE.with_alpha(0.5));
        ctx.glyph_run(&font)
            .font_size(font_size)
            .hint(true)
            .fill_glyphs(glyphs.into_iter());
    });
```

This was fixed by modifying the `record` trait such that it is no longer
built in. Now it eagerly pushes a `set_transform` such that the
transform of the recording is tracked. This correctly builds the glyphs.

Fix: Recording is now correctly absolutely positioned when it is
replayed because it _always_ pushes a set_transform on recording start.

### Test plan

Added multiple unique test cases and also checked the examples.


Results from `cargo run -p vello_hybrid_winit --release`:

```sh
# GhostTiger
SVG  | Fresh     : 12.040ms | Strips: 6850 | Alphas: 308048
SVG  | Identical : 0.957ms | Strips: 6850 | Alphas: 308048
SVG  | Identical : 0.483ms | Strips: 6850 | Alphas: 308048
# Text
Text | Fresh     : 0.693ms | Strips: 89 | Alphas: 2896
Text | Fresh     : 0.215ms | Strips: 89 | Alphas: 2896
Text | Fresh     : 0.212ms | Strips: 91 | Alphas: 3248
Text | Identical : 0.004ms | Strips: 91 | Alphas: 3248
Text | Identical : 0.015ms | Strips: 91 | Alphas: 3248
```

Results from `cargo run_wasm -p native_webgl --release`:
- Checked performance tab and you can't even see the
`request_animation_frame` that contains the cached render 😆 .

### Caveats

Note – recording API still crashes if the sparse strips go off the top
left of the screen. That behavior remains consistent with `main`.

This PR fixes the transform handling of the recording – maintaining the
expected behavior that the recording will be replayed at the transforms
that it was recorded at. In the future we can extend this API to add a
translation when we support it. Or we can add a new API that executes a
recording at a translation.