[sparse strips] Remove `GlyphRenderer` dead code in tests (#1241)
The `Renderer` trait for sparse strips testing _does not_ actually
depend on `GlyphRenderer` because `glyph_run` is calling the underlying
implementation on either the `RenderContext` (Vello CPU) or `Scene`
(Vello Hybrid) directly.

### Test plan

This is a test-only change, and can be validated by ensuring that this
compiles & tests continue to pass.