Introduce a new clipping algorithm for non-layer based clipping (#1203)
# Motivation
Currently, the only way to do efficient clipping in Vello Sparse is by
pushing a new isolated layer and associating a clip path with that
layer. This might be sufficient for cases like for example SVG, but the
problem is that doing so forces the user to conflate layers and
clipping, which is not what should be done in some imaging models. For
example, clip paths in PDF have nothing to do with isolation, and it is
therefore desirable to be able to separate the two. I believe the same
is the case for HTML canvas, where the clip stack is not really related
to layers, and the fact they are causes problems (see
https://github.com/linebender/vello/issues/1117). In practice, this
mostly doesn't cause issues as (to my understanding) source-over
compositing with premultiplied alpha is associative and it therefore
mostly doesn't matter whether you clip with isolation or not. However,
problems start to occur once you add masks or blending to the mix, as
the behavior with and without isolation is not identical.

It should be noted that above can be simulated by creating a new mask
for each clip path, but this is obviously expensive, especially for
deeply-nested clips.

# Solution
Because of this, I propose a second way to support clipping in Vello CPU
(could be added to hybrid as well) that is based on an independent clip
stack that can be pushed and popped to without actually having to create
new layers. In particular, you do not need to clear the clip stack
before rasterizing, which should adddress the issue in #1117.

I think how this works is very neat and demonstrates how useful sparse
strips are! Fundamentally, this is based on a new algorithm that, given
two paths in sparse strips representation, computes a new path in sparse
strips representation that represents the intersection of the two. While
the actual implementation is a bit finicky, conceptually it's very
simple: We iterate over each stripped and fill region of the two paths
in lock step and determine all of the overlaps between the two. In case
we have two fill regions, the overlap region will also be filled. In
case we have one strip and one fill region, the overlap will copy the
alpha mask of the strip region. Finally, if we have two strip regions,
we combine the alpha masks of both. That's it, this is enough to compute
a new sparse strips path that when filled is equivalent to intersecting
the two strips. The beauty of this is that this is incredibly fast
because of the usual sparseness of paths and the result will also still
be as sparse as possible as we preserve sparse fill regions whenever
possible.

Given this, we introduce two new `push_clip_path` and `pop_clip_path`
methods that operate on their own clip stack and completely
independently from the push/pop layer commands. In case a clip path is
currently active, once we draw a path we simply calculate the
intersection after converting it into sparse strips, and then use that
new path as the basis in coarse rasterization instead. In case we call
`push_clip_path` multiple times, we can simply intersect the new clip
path with the one in the previous stack position, to end up with one
single new clip path that represents the intersection of all previous
ones. Therefore, even for deeply nested clips we only need to perform
one intersection per path drawn.

The disadvantage is that in multi-threading, it requires us to process
new clip paths on the main thread, but the advantage is that (unlike the
current clip algorithm) causes no additional overhead during
rasterization (which is currently the only serial part), and the
clipping calculation is done by each worker separately for the path they
are currently processing, which means it scales very well!

# Testing
I will add more tests, but I confirmed the current version to cause no
regressions in my PDF test suite.