Add tile intersection checking (#1293)
This PR adds generation of intersection data to `make_tiles.` 

For our MSAA calculations, we need the points where a line intersects
with a tile. Since the end goal is to perform the MSAA in parallel, we
need these intersections to be watertight between different threads. By
performing these calculations in `make_tiles`, we can create a source of
ground truth, ensuring watertightness.

Although it is feasible to calculate the exact intersection points in
`make_tiles`, we defer that work to the gpu. Instead we create an
intersection bitmask, which unambiguously defines which edges the line
intersects. We say a line *intersects* the edge of a tile if it touches
that edge AND continues into another tile. A endpoint that exactly
touches an edge does NOT count as intersecting, though it may still
contribute to winding (in the case of a top edge touch).

The bitmask is 5 bits, and will be consumed downstream by the
rasterization stage:
`P(erfect) | R(ight) | L(eft) | B(ottom) | T(op)`
The lower 4 bits correspond to the intersection edges. The "Perfect" bit
is necessary to resolve ambiguity in cases where a line perfectly
intersects with a corner of a tile.

---
Cases of 3 bit ambiguity can be resolved by always calculating
intersections on the opposing edges of the tile.
Consider a tile with intersections `T | L | B`:
```
o--------+
|\       |
| \      |
|  \     | 
+---o----+
```
Calculate the intersections on `T` and `B`!

---

Cases of 2 bit ambiguity require the "Perfect" bit, which is set when
there is exactly one unique edge intersection.
Consider a tile with intersections `T | L`:

This is valid:
```
o--------+
|\       |
| o      |
|        | 
+--------+
```

But so is this:
 ```
+--o-----+
| /      |
o        |
|        | 
+--------+
```

With the perfect bit, the first case would be `P | T | L` and the second case would be `T | L`

---------

Co-authored-by: Thomas Smith <thomsmit@google.com>